refactor scripts to allow further options via ini
this commit adds dx_config.ini (_default in the repo)
This file lays out locations for the 3 emulators, as well as paths needed for base games
This allows the user to optionally include this information, and quickly build, and run the game with one click. for rpcs3, you can even include the option to auto start discord rich presence. There is also an option to build to and launch debug wii as well