move dx to pal eboot with special sauce
this eboot is the pal eboot, to match in parity to rb3e, even tho it doesnt matter at this time, as rb3e uses its own eboot, it will allow future patches coming from rb3e to be applied easier to dx, if needed.

This also brings with it the song blacklist patch, that DX never had on ps3, meaning hierkommtalex and rockandrollstar now work in their original format
This eboot also has a special rb3e feature baked in. The ability to play unencrypted midis on hardware.