brutal themed song select
This commit adds a call to the setlist UITrigger in song_select_panel to trigger the red theme, typically used for setlist creation. This helps visually cue the player that they are indeed, playing brutal mode. Rock Band 4 has a similar red styling when using brutal mode.

I simply added to the existing if_else chains that determine which trigger to fire to accomplish this. As well, i added to the existing .lbl if_else chain to rename the title to say "Brutal Mode" tying it all together.
There is also an addition in os_milo to fire UITriggers, previously only EventTriggers were hooked up