Apply RB4 RBN updates
Apply updated from RB4 RBN releases into RB3DX
Includes midi updates, and metadata updates (difficulty and genre changes)
Does NOT include section names, as the game does not read those properly (the percentage sections from the base midis are used in the selection practice section screen, but the updated markers are used in other places)
The For the Love of God (Live) BRE is added, as the game does read the EVENTS when the track is loaded
In the case of pre-existing update files some cases, if the reported difference was negligible (e.g. a note is one tick off) I just left the existing update file alone.
Stuff like harmonies (that weren't used in the RB4 versions) and pro keys/strings upgrades have been taken into account