Add tbrb/gdrb loading phrases/broken quotes
As it turns out, if the \q markers appears directly before or after a double quote, the songs_updates.dta file can't process it, so I've removed all of the \q that come at the very start or end of loading phrases by either turning them into single quotes or placing them before periods. (If a \q appears in the middle of a loading phrase, it can stay.)

Also, adding loading phrases from TBRB and adding new ones to GDRB. TBRB phrases are taken from the story photos, GDRB ones are of my own creation.