Revert most RB4 RBN changes
Revert most of the changes that were made to RBN charts in RB4. Many of them were akin to complete overhauls, as such it's better that the original charts be playable instead.
Tom marker fixes are intact, in many cases the marker placement was improved over previous upgrades
A few lane markers kept if it was feasible to include them
Removed  "blue and orange on Easy/Medium" updates that were done to earlier RBN songs.
Care was taken to make sure existing Harmonies/Pro Strings/Pro keys upgrades were kept intact
In This Moment - Mechanical Love harms upgrade retained for now, even though the PART VOCALS chart is reverted

This is made a draft for now as I still need want some feedback as to what to keep and such, and some time to fix errors and such.